[Fleet] fix error when clearing upgrade input (#175198)
Closes https://github.com/elastic/kibana/issues/175186

Fixed UI error in the custom upgrade version input.
To verify:
- enroll an agent with version 8.13-snapshot
- click upgrade agent
- type anything and then clear the input
- the UI should not move to error state

Fixed another bug when I tested to upgrade to a newer build version
(newer patch), and got an API error saying can't upgrade to a newer
version than kibana version.
To verify:
- click upgrade agent, type `8.13.1+build123456789`, submit
- verify that the API returns 200

https://github.com/elastic/kibana/assets/90178898/80b4e12e-3aad-41b7-b39d-8d28d6253772

- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios
